Update script.js
completed negative partially and edited fnum and snum to strings

+++ script.js
@@ -15,15 +15,17 @@ function divide(a, b) {
 }
 
 function operate(fnum, snum, operator) {
+  fnum = Number(fnum);
+  snum = Number(snum);
   switch (operator) {
     case "+":
-      return add(fnum, snum);
+      return String(add(fnum, snum));
     case "-":
-      return subtract(fnum, snum);
+      return String(subtract(fnum, snum));
     case "*":
-      return multiply(fnum, snum);
+      return String(multiply(fnum, snum));
     case "/":
-      return Math.round(divide(fnum, snum) * 1000) / 1000;
+      return String(Math.round(divide(fnum, snum) * 1000) / 1000);
     default:
       alert("Invalid operator!");
   }
@@ -33,8 +35,7 @@ let fnum = null;
 let operator = null;
 let snum = null;
 let displayValue = null;
-let isDecimal = false;
-let isNegative = false;
+let justEvaluated = false;
 let operatorConvertor = {
   "+": "+",
   "-": "-",
@@ -44,84 +45,59 @@ let operatorConvertor = {
 
 function handleNumber(input) {
   // if input is a number, check if operator exists, if so concat it to snum, else concat it to fnum
-  // it doesn't matter if the number is just derived, I like it this way
+  // if number is just evaluated, rewrite over it
   if (operator === null) {
-    if (fnum === null) {
-      if (isDecimal) {
-        fnum = input / 10;
-      } else if (isNegative) {
-        fnum = -input;
-        isNegative = false;
-      } else {
-        fnum = input;
-      }
+    if (justEvaluated) {
+      fnum = String(input)
+      justEvaluated = false;
     } else {
-      if (isDecimal) {
-        fnum = Number(String(fnum) + "." + String(input));
-        isDecimal = false;
-      } else {
-        fnum = Number(String(fnum) + String(input));
-      }
+      fnum += String(input);
     }
-    displayValue = String(fnum);
+    displayValue = fnum;
   } else {
-    if (snum === null) {
-      if (isDecimal) {
-        snum = input / 10;
-      } else if (isNegative) {
-        snum = -input;
-        isNegative = false;
-      } else {
-        snum = input;
-      }
-    } else {
-      if (isDecimal) {
-        snum = Number(String(snum) + "." + String(input));
-        isDecimal = false;
-      } else {
-        snum = Number(String(snum) + String(input));
-      }
-    }
-    displayValue = String(fnum) + operatorConvertor[operator] + String(snum);
+    snum += String(input);
+    displayValue = fnum + operatorConvertor[operator] + snum;
   }
 }
 
 function handleOperator(input) {
   // if input is an operator, check if snum exists, if so perfrom the previous calculation first, else update operator
-  // if - is given after a num, treat it as minus sign, if given when fnum is null or after a non - operator, treat it as negative sign
-  // if given after snum, treat it as minus sign and a calculate previous nums first
+  // if - is given after a num, treat it as minus sign, and if given after snum, calculate previous nums first
+  // if given when fnum is null or after a non minus operator, treat it as negative sign
+  justEvaluated = false;
   if (snum === null) {
     if (input === "-") {
       if (fnum === null) {
-        isNegative = true;
-        displayValue = "-";
+        fnum = "-"
+        displayValue = fnum;
       } else if (operator === null) {
         operator = input;
-        displayValue = String(fnum) + operatorConvertor[operator];
+        displayValue = fnum + operatorConvertor[operator];
       } else if (operator !== null) {
-        isNegative = true;
-        displayValue = String(fnum) + operatorConvertor[operator] + "-";
+        snum = "-"
+        displayValue = fnum + operatorConvertor[operator] + snum;
       }
      } else {
       operator = input;
       if (fnum === null) {
-        fnum = 0;
-        displayValue = "0" + operatorConvertor[operator];
-      } else {
-        displayValue = String(fnum) + operatorConvertor[operator];
-      }
-    }
+        fnum = "0";
+      }  
+      displayValue = fnum + operatorConvertor[operator];
+     }
   } else {    // if snum is not null, operator is not null too, thus fnum is not null too
-    if (fnum === 0 && operator === "/") {
+    if ((Number(snum) === 0 || Number(snum) === -0) && operator === "/") {
       handleClear();
       displayValue = "ERROR";
     } else {
       fnum = operate(fnum, snum, operator);
-      operator = input;
-      snum = null;
-      isDecimal = false;
-      isNegative = false;
-      displayValue = String(fnum) + operatorConvertor[operator];
+      if (isNan(fnum)) {
+        handleClear();
+        displayValue = "ERROR";
+      } else {
+        operator = input;
+        snum = null;
+        displayValue = fnum + operatorConvertor[operator];
+      }
     }
   }
 }
@@ -132,88 +108,66 @@ function handleClear() {
   operator = null;
   snum = null;
   displayValue = null;
-  isDecimal = false;
-  isNegative = false;
+  justEvaluated = false;
 }
 
 function handleDelete() {
   // if input is delete, remove the last input (only limited to before = is pressed)
-  if (isNegative) {
-    displayValue = displayValue.slice(0, displayValue.length - 1);
-    isNegative = false;
-  } else if (fnum !== null) {
-    if (operator === null) {
-      if (String(fnum).length <= 1) {
-        fnum = null;
-      } else if (fnum.includes("-") && String(fnum).length === 2) {
-        pass;
-      } if (fnum.includes("-") && String(fnum).length === 2) {
-        fnum = Number(String(fnum).slice(0, String(fnum).length - 1));
-      }
-    } else if (snum === null) {
-      operator = null;
+  justEvaluated = false;
+  if (operator === null) {      // dealing with fnum
+    if (fnum.length === 1) {
+      fnum = null;
     } else {
-      if (String(snum).length <= 1) {
-        snum = null;
-      } else {
-        snum = Number(String(snum).slice(0, String(snum).length - 1));
-      }
+      fnum = fnum.slice(0, fnum.length - 1);
     }
-    displayValue = displayValue.slice(0, displayValue.length - 1);
+  } else if (snum === null) {    // dealing with operator
+    operator = null;
+  } else {        // dealing with snum
+    if (snum.length === 1) {
+      snum = null;
+    } else {
+      snum = snum.slice(0, snum.length - 1);
+    }
   }
+  displayValue = displayValue.slice(0, displayValue.length - 1);
 }
 
 function handleEvaluate() {
   // if input is evaluate, calculate based on current values
   // if fnum is present but snum is absent, perfrom the operator on fnum
   // if operator is missing too or no values provided at all, do nothing
-  if (fnum !== null && operator !== null && snum !== null) {
-        if (fnum === 0 && operator === "/") {
-          fnum = null;
-          operator = null;
-          snum = null;
-          isDecimal = false;
-          isNegative = false;
-          displayValue = "ERROR";
-        } else {
-          fnum = operate(fnum, snum, operator);
-          snum = null;
-          operator = null;
-          isDecimal = false;
-          isNegative = false;
-          displayValue = String(fnum);
-        }
-      } else if (operator !== null && snum === null) {  // if operator is not null, fnum cannot be null too
-        if (fnum === 0 && operator === "/") {
-          handleClear();
-          displayValue = "ERROR";
-        } else {
-          if (isDecimal) {
-            snum = 0;
-          } else {
-            snum = fnum;
-          }
-          fnum = operate(fnum, snum, operator);  
-          operator = null;
-          snum = null;
-          displayValue = String(fnum);
-        }
-      }
-      isDecimal = false;
+  if (snum === null) {
+    snum = fnum;
+  }
+  if (snum === 0 && operator === "/") {
+    handleClear();
+    displayValue = "ERROR";
+  }
+  fnum = operate(fnum, snum, operator);
+  if (isNan(fnum)) {
+    handleClear();
+    displayValue = "ERROR";
+  } else {
+    snum = null;
+    operator = null;
+    justEvaluated = true;
+    displayValue = fnum;
+  }
 }
 
 function handleDecimal() {
   // if input is decimal, add decimal to number
-  // if no number present, set counter to true such that next number type would cause
-  if (fnum === null) {
-    fnum = 0;
-    isDecimal = true;
-    displayValue = "0.";
-  } else if (operator === null) {
-    if (!String(fnum).includes(".")) {
-      isDecimal = true;
-      displayValue = String(fnum) + ".";
+  if (fnum === null || justEvaluated) {
+    fnum = "0.";
+    justEvaluated = false;
+    displayValue = fnum;
+  } else if (operator === null && !fnum.includes(".")) {
+    if (fnum === "-") {
+      fnum = "-0.";
+    } else {
+      fnum += "."
     }
+    displayValue = fnum;
   } else if (snum === null) {
     snum = 0;
     isDecimal = true;
@@ -254,21 +208,17 @@ btns.forEach(btn => {
   btn.addEventListener("click", e => {
     let input = buttonLabel[e.target.getAttribute("class")];
     if (displayValue != null && displayValue.length >= 13 && input !== undefined) {
+      handleClear();
       displayValue = "MAX LENGTH";
-      fnum = null;
-      operator = null;
-      snum = null;
-      displayValue = null;
-      isDecimal = false;
     } else if (!isNaN(input)) {
       handleNumber(input)
     } else if ("+-*/".includes(input)) {
       handleOperator(input);
     } else if (input === "clear") {
       handleClear();
-    } else if (input === "delete") {
+    } else if (input === "delete" && fnum !== null) {
       handleDelete();
-    } else if (input === "=") {
+    } else if (input === "=" && fnum !== null && operator !== null) {
       handleEvaluate();
     } else if (input === ".") {
       handleDecimal();
